+++ nondefaced_detector/models/model.py
